testing and some missed edge cases
